requirements.txt
bcrypt==3.1.7
boto3==1.10.16
botocore==1.13.16
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Django==2.2.7
django-cors-headers==3.1.1
docutils==0.15.2
idna==2.8
jmespath==0.9.4
mysql-connector-python==8.0.18
protobuf==3.10.0
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.6
